Add UserAuthController and update routes
This commit introduces the UserAuthController, with the methods 'getUser' and 'createUser' for handling user authentication. Changes in the routes file include the addition of '/user' route using these new methods and commenting out the previous '/user' route. Jetstream features and the APP_URL in the .env.example file have also been updated accordingly.